优化Homepage源码结构
