+++ app/src/main/java/com/yenaly/han1meviewer/ui/navigation/main/HomeRoute.kt
@@ -17,7 +17,7 @@ import com.yenaly.han1meviewer.ui.activity.MainActivity
 import com.yenaly.han1meviewer.ui.component.AnnouncementDialog
 import com.yenaly.han1meviewer.ui.component.TripleButtonDialog
 import com.yenaly.han1meviewer.ui.screen.home.HomePageScreen
-import com.yenaly.han1meviewer.ui.screen.home.LocalSearchHistoryQuery
+import com.yenaly.han1meviewer.ui.screen.home.homepage.LocalSearchHistoryQuery
 import com.yenaly.han1meviewer.ui.viewmodel.CheckInCalendarViewModel
 import com.yenaly.yenaly_libs.utils.copyTextToClipboard
 import com.yenaly.yenaly_libs.utils.showShortToast
