+++ app/src/main/java/com/yenaly/han1meviewer/ui/screen/home/homepage/HomePageUtils.kt
@@ -1,0 +1,7 @@
+package com.yenaly.han1meviewer.ui.screen.home.homepage
+
+internal fun HomeCategory.toAdvancedSearchParams(): Map<String, String> = buildMap {
+    genre?.let { put("genre", it) }
+    sort?.let { put("sort", it) }
+    tags?.let { put("tags", it) }
+}
